Enabled automatic alarm Added support for 	- populating EVENT row 	- removing events from Event row if a threshold is reached 	- removing RDR events (only those are marked for deletion - via the RPT populate_rpt call) 	- removing RTP events

+++ include/saHpiRdrTable.h
@@ -90,7 +90,7 @@ extern          "C" {
 		  oid *rpt_oid, size_t rpt_oid_len,
 		  oid *resource_oid, size_t resource_oid_len);
 
-  int purge_rdr( void );
+  unsigned long  purge_rdr( void );
 int
 delete_rdr_row(SaHpiDomainIdT domain_id,
 	       SaHpiResourceIdT resource_id,
